LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-history-entry/opaque-origin-expected.txt



FAIL navigation.currentEntry/entries()/canGoBack/canGoForward in an opaque origin iframe assert_equals: expected null but got object "[object NavigationHistoryEntry]"
